LPS-130175 Move CollectionActiveItemContext component
