+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/ItemConfiguration.js
@@ -17,12 +17,12 @@ import ClayTabs from '@clayui/tabs';
 import PropTypes from 'prop-types';
 import React, {useEffect, useMemo, useState} from 'react';
 
-import {useCollectionActiveItemContext} from '../../../../../app/components/CollectionActiveItemContext';
 import {CollectionItemContext} from '../../../../../app/components/CollectionItemContext';
 import {
 	useActiveItemId,
 	useActiveItemType,
 } from '../../../../../app/components/Controls';
+import {useCollectionActiveItemContext} from '../../../../../app/contexts/CollectionActiveItemContext';
 import {useSelectorCallback} from '../../../../../app/store/index';
 import {deepEqual} from '../../../../../app/utils/checkDeepEqual';
 import {useId} from '../../../../../app/utils/useId';
